+++ src/Input/Input.stories.tsx
@@ -1,28 +1,36 @@
-import React, { useState } from "react"
-import { withKnobs, number, boolean, text, select, color } from "@storybook/addon-knobs"
+import React, { useState } from 'react'
+import {
+  withKnobs,
+  number,
+  boolean,
+  text,
+  select,
+  color,
+} from '@storybook/addon-knobs'
 
-import Input from "./Input"
-import { Meta } from "@storybook/react/types-6-0"
+import Input from './Input'
+import { Meta } from '@storybook/react/types-6-0'
 
 export default {
   component: Input,
-  title: "Components/Input",
+  title: 'Components/Input',
   decorators: [withKnobs],
 } as Meta
 
 export function Index() {
-  const [value, setValue] = useState<string>("")
+  const [value, setValue] = useState<string>('')
 
-  const width = number("width", 300)
-  const placeholder = text("placeholder", "텍스트를 입력해주세요.")
-  const errorMessage = text("errorMessage", "에러메시지")
-  const disabled = boolean("disabled", false)
-  const useCancleButton = boolean("useCancleButton", false)
-  const isError = boolean("isError", false)
-  const mode = select("mode", ["line", "basic"], "line")
-  const readOnly = boolean("readOnly", false)
-  const tabIndex = number("tabIndex", 10)
-  const borderColor = color("borderColor", "#FAC609")
+  const width = number('width', 300)
+  const placeholder = text('placeholder', '텍스트를 입력해주세요.')
+  const errorMessage = text('errorMessage', '에러메시지')
+  const disabled = boolean('disabled', false)
+  const useCancleButton = boolean('useCancleButton', false)
+  const isError = boolean('isError', false)
+  const mode = select('mode', ['line', 'basic'], 'line')
+  const readOnly = boolean('readOnly', false)
+  const tabIndex = number('tabIndex', 10)
+  const borderColor = color('borderColor', '#FAC609')
+  const useCursor = boolean('useCursor', true)
   return (
     <Input
       mode={mode}
@@ -37,21 +45,22 @@ export function Index() {
       readOnly={readOnly}
       tabIndex={tabIndex}
       onFocus={() => {
-        console.log("onFocus")
+        console.log('onFocus')
       }}
       onClick={() => {
-        console.log("onClick")
+        console.log('onClick')
       }}
       onBlur={() => {
-        console.log("onBlur")
+        console.log('onBlur')
       }}
       onClickCancelButton={() => {
-        console.log("onClickCancelButton")
+        console.log('onClickCancelButton')
       }}
       onKeyDown={() => {
-        console.log("onKeyDown")
+        console.log('onKeyDown')
       }}
       borderColor={borderColor}
+      useCursor={useCursor}
     />
   )
 }