+++ src/Input/Input.tsx
@@ -1,19 +1,19 @@
-import React from 'react'
-import styled from 'styled-components'
+import React from "react"
+import styled from "styled-components"
 
-import Icon from '../Icon/Icon'
+import Icon from "../Icon/Icon"
 
 const InputContainer = styled.div``
 const InputBox = styled.div<{
   width: number
   disabled: boolean
-  mode: 'line' | 'basic'
+  mode: "line" | "basic"
   borderColor: string
 }>`
   padding: 6px ${props => props.width * 0.04}px;
   ${props =>
     `${
-      props.mode == 'line'
+      props.mode == "line"
         ? `
                 border-bottom: 1px solid #dfdedd;
             `
@@ -23,38 +23,38 @@ const InputBox = styled.div<{
     }`}
   &:hover {
     ${props =>
-      !props.disabled
-        ? props.mode == 'line'
-          ? `border-bottom: 1px solid ${
-              props.disabled ? '#dfdedd' : props.borderColor
-            }`
-          : `border: 1px solid ${
-              props.disabled ? '#dfdedd' : props.borderColor
-            }`
-        : ''}
+    !props.disabled
+      ? props.mode == "line"
+        ? `border-bottom: 1px solid ${
+          props.disabled ? "#dfdedd" : props.borderColor
+        }`
+        : `border: 1px solid ${
+          props.disabled ? "#dfdedd" : props.borderColor
+        }`
+      : ""}
   }
   &:focus-within {
     ${props =>
-      !props.disabled
-        ? props.mode == 'line'
-          ? `border-bottom: 1px solid ${
-              props.disabled ? '#dfdedd' : props.borderColor
-            }`
-          : `border: 1px solid ${
-              props.disabled ? '#dfdedd' : props.borderColor
-            }`
-        : ''}
+    !props.disabled
+      ? props.mode == "line"
+        ? `border-bottom: 1px solid ${
+          props.disabled ? "#dfdedd" : props.borderColor
+        }`
+        : `border: 1px solid ${
+          props.disabled ? "#dfdedd" : props.borderColor
+        }`
+      : ""}
   }
   display: flex;
   align-items: center;
   width: ${props => props.width}px;
-  border-radius: ${props => (props.mode == 'line' ? '0px' : '3px')};
+  border-radius: ${props => (props.mode == "line" ? "0px" : "3px")};
   justify-content: space-between;
   ${props =>
     `${
-      props.mode == 'line'
-        ? props.disabled && 'border-bottom: 1px dashed #dfdedd;'
-        : props.disabled && 'background-color: #F0F0F0;'
+      props.mode == "line"
+        ? props.disabled && "border-bottom: 1px dashed #dfdedd;"
+        : props.disabled && "background-color: #F0F0F0;"
     }`}
 `
 const InputElement = styled.input<{ width: number }>`
@@ -71,7 +71,7 @@ const SubText = styled.p`
 `
 
 export type InputType = {
-  mode: 'line' | 'basic'
+  mode: "line" | "basic"
   placeholder: string
   value: string
   onChange: (value: string) => void
@@ -144,6 +144,7 @@ function Input({
             width={20}
             color="#DFDEDD"
             onClick={onClickCancelButton}
+            useCursor={true}
           />
         )}
       </InputBox>